리뷰, 가게 정보 마이그레이션

+++ app/src/main/java/com/zion830/threedollars/repository/model/v2/request/EditReviewRequest.kt
@@ -7,5 +7,5 @@ data class EditReviewRequest(
     @SerializedName("contents")
     val contents: String = "",
     @SerializedName("rating")
-    val rating: Int = 0
+    val rating: Float = 0f
 )
