+++ app/src/main/java/com/zion830/threedollars/repository/model/v2/response/NewReviewResponse.kt
@@ -18,7 +18,7 @@ data class NewReview(
     @SerializedName("createdAt")
     val createdAt: String = "",
     @SerializedName("rating")
-    val rating: Int = 0,
+    val rating: Float = 0f,
     @SerializedName("reviewId")
     val reviewId: Int = 0,
     @SerializedName("storeId")
