+++ app/src/main/java/com/zion830/threedollars/repository/model/v2/response/store/StoreList.kt
@@ -11,11 +11,11 @@ data class StoreList(
     @SerializedName("distance")
     val distance: Int = 0,
     @SerializedName("latitude")
-    val latitude: Int = 0,
+    val latitude: Double = 0.0,
     @SerializedName("longitude")
-    val longitude: Int = 0,
+    val longitude: Double = 0.0,
     @SerializedName("rating")
-    val rating: Int = 0,
+    val rating: Float = 0f,
     @SerializedName("storeId")
     val storeId: Int = 0,
     @SerializedName("storeName")
